Update api with ID validation

+++ app/api/tasks/[id]/route.ts
@@ -5,25 +5,37 @@ import { NextRequest, NextResponse } from "next/server";
 export async function DELETE(
     request: NextRequest, 
     { params }: { params: { id: string }}) {
-    const deletedTask = await prisma.task.delete({
-        where: { id: Number(params.id) },
+
+    const task = await prisma.task.findUnique({
+        where: { id: parseInt(params.id) }
+    });
+    if (!task)
+        return NextResponse.json({ error: 'Invalid task' }, { status: 404 });
+    
+    await prisma.task.delete({
+        where: { id: task.id },
     });
 
-    return NextResponse.json(deletedTask, { status: 200 })
+    return NextResponse.json({});
 }
 
 export async function PATCH(
     request: NextRequest, 
     { params }: { params: { id: string }}) {
+        
     const body = await request.json();
     const validation = updateTaskSchema.safeParse(body);
     if (!validation.success)
         return NextResponse.json(validation.error.format(), { status: 400 });
 
+    const task = await prisma.task.findUnique({
+        where: { id: parseInt(params.id) }
+    });
+    if (!task)
+        return NextResponse.json({ error: 'Invalid task' }, { status: 404 });
+
     const updatedTask = await prisma.task.update({
-        where: {
-          id: parseInt(params.id)
-        },
+        where: { id: task.id },
         data: {
             completed: body.completed, 
             title: body.title, 
@@ -33,5 +45,5 @@ export async function PATCH(
         }
       });
 
-    return NextResponse.json(updatedTask, { status: 200 });  
+    return NextResponse.json(updatedTask);  
 }
